+++ packages/gamut-storybook/stories/Button-story.js
@@ -7,7 +7,7 @@ import ButtonBase from '@codecademy/gamut/ButtonBase';
 import Spinner from '@codecademy/gamut/Spinner';
 import RadialProgress from '@codecademy/gamut/RadialProgress';
 import { select, text, boolean } from '@storybook/addon-knobs';
-import gamut from '@codecademy/gamut-styles/variables';
+import { colors } from '@codecademy/gamut-styles/variables';
 
 import { addonInfoOptions as options } from './options';
 
@@ -32,33 +32,84 @@ themeKeys.forEach(k => {
 });
 
 const btnStyle = {
-  margin: '0.5rem',
+  marginRight: '1rem',
+  marginBottom: '1rem',
 };
 
 storiesOf('Component/Button', module)
   .add('About Buttons', () => <div>Button philosophy goes here.</div>)
   .add(
-    'Standard Button themes',
+    'All Button Themes',
     withInfo({
       ...options,
     })(() => (
       <div>
-        <Button style={btnStyle}>{text('Label', 'Submit')}</Button>
-        <Button style={btnStyle} theme="primary" href="#">
-          Primary (red)
-        </Button>
-        <Button style={btnStyle} theme="secondary">
-          Secondary (mint)
-        </Button>
-        <Button style={btnStyle} theme="blue" href="#">
-          Blue
-        </Button>
-        <Button style={btnStyle} theme="yellow">
-          Yellow
-        </Button>
-        <Button style={btnStyle} theme="white" href="#">
-          White
-        </Button>
+        <div style={{ padding: '0.5rem' }}>
+          <h3>Standard Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button key={`${theme}-onlight`} style={btnStyle} theme={theme}>
+              {theme}
+            </Button>
+          ))}
+          <h3>Outline Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button
+              key={`${theme}-onlight-outline`}
+              style={btnStyle}
+              theme={theme}
+              outline
+            >
+              {theme}
+            </Button>
+          ))}
+          <h3>Flat Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button
+              key={`${theme}-onlight-flat`}
+              style={btnStyle}
+              theme={theme}
+              flat
+            >
+              {theme}
+            </Button>
+          ))}
+        </div>
+        <div
+          style={{
+            padding: '0.5rem',
+            background: colors.swatches.grey[900],
+            color: 'white',
+          }}
+        >
+          <h3>Standard Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button key={`${theme}-ondark`} style={btnStyle} theme={theme}>
+              {theme}
+            </Button>
+          ))}
+          <h3>Outline Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button
+              key={`${theme}-ondark-outline`}
+              style={btnStyle}
+              theme={theme}
+              outline
+            >
+              {theme}
+            </Button>
+          ))}
+          <h3>Flat Buttons</h3>
+          {themeKeys.map(theme => (
+            <Button
+              key={`${theme}-ondark-flat`}
+              style={btnStyle}
+              theme={theme}
+              flat
+            >
+              {theme}
+            </Button>
+          ))}
+        </div>
       </div>
     ))
   )
@@ -123,7 +174,7 @@ storiesOf('Component/Button', module)
     })(() => (
       <div
         style={{
-          background: gamut.colors.portal.midnightblue,
+          background: colors.portal.midnightblue,
           padding: '0.5rem',
         }}
       >
@@ -176,5 +227,35 @@ storiesOf('Component/Button', module)
       >
         {text('Label', 'Submit')}
       </Button>
+    ))
+  )
+  .add(
+    'New',
+    withInfo({
+      text: 'New Styles',
+      ...options,
+    })(() => (
+      <div>
+        <Button style={btnStyle} theme="yellow">
+          Super
+        </Button>
+        <Button style={btnStyle} theme="blue">
+          Primary (new)
+        </Button>
+        <Button style={btnStyle} theme="blue" outline>
+          Seconary (new)
+        </Button>
+        <Button style={btnStyle} theme="blue" flat>
+          Flat
+        </Button>
+
+        <p>
+          This is an example of a new{' '}
+          <Button theme="grey" link href="#">
+            Link
+          </Button>{' '}
+          style button.
+        </p>
+      </div>
     ))
   );