+++ packages/gamut/CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.6.20"></a>
+## [1.6.20](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.19...@codecademy/gamut@1.6.20) (2018-03-22)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.19"></a>
+## [1.6.19](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.18...@codecademy/gamut@1.6.19) (2018-03-20)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.18"></a>
+## [1.6.18](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.17...@codecademy/gamut@1.6.18) (2018-03-13)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.17"></a>
+## [1.6.17](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.16...@codecademy/gamut@1.6.17) (2018-03-07)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.16"></a>
+## [1.6.16](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.15...@codecademy/gamut@1.6.16) (2018-03-07)
+
+
+### Bug Fixes
+
+* **gamut:** remove button underline on hover ([bccd7d8](https://github.com/RyzacInc/gamut/commit/bccd7d8))
+
+
+
+
+<a name="1.6.15"></a>
+## [1.6.15](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.14...@codecademy/gamut@1.6.15) (2018-03-07)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.14"></a>
+## [1.6.14](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.13...@codecademy/gamut@1.6.14) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.13"></a>
+## [1.6.13](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.12...@codecademy/gamut@1.6.13) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.12"></a>
+## [1.6.12](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.11...@codecademy/gamut@1.6.12) (2018-03-06)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
+<a name="1.6.11"></a>
+## [1.6.11](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.10...@codecademy/gamut@1.6.11) (2018-03-02)
+
+
+
+
+**Note:** Version bump only for package @codecademy/gamut
+
 <a name="1.6.10"></a>
 ## [1.6.10](https://github.com/RyzacInc/gamut/compare/@codecademy/gamut@1.6.9...@codecademy/gamut@1.6.10) (2018-02-21)